+++ src/gui/MainWindow.h
@@ -27,8 +27,17 @@
 
 #include <kImageAnnotator/KImageAnnotator.h>
 
+#include "src/gui/TrayIcon.h"
+#include "src/gui/ClipboardWrapper.h"
 #include "src/gui/aboutDialog/AboutDialog.h"
 #include "src/gui/settingsDialog/SettingsDialog.h"
+#include "src/gui/operations/AddWatermarkOperation.h"
+#include "src/gui/operations/UploadOperation.h"
+#include "src/gui/operations/HandleUploadResultOperation.h"
+#include "src/gui/globalHotKeys/GlobalHotKeyHandler.h"
+#include "src/gui/captureHandler/CaptureHandlerFactory.h"
+#include "src/gui/captureHandler/ICaptureChangeListener.h"
+#include "src/gui/widgetHider/WidgetHiderFactory.h"
 #include "src/widgets/MainToolBar.h"
 #include "src/backend/imageGrabber/AbstractImageGrabber.h"
 #include "src/backend/config/KsnipConfigProvider.h"
@@ -39,14 +48,6 @@
 #include "src/common/provider/ApplicationTitleProvider.h"
 #include "src/common/dtos/CaptureFromFileDto.h"
 #include "src/common/handler/DragAndDropHandler.h"
-#include "src/gui/operations/AddWatermarkOperation.h"
-#include "src/gui/operations/UploadOperation.h"
-#include "src/gui/operations/HandleUploadResultOperation.h"
-#include "src/gui/globalHotKeys/GlobalHotKeyHandler.h"
-#include "src/gui/TrayIcon.h"
-#include "src/gui/ClipboardWrapper.h"
-#include "src/gui/captureHandler/CaptureHandlerFactory.h"
-#include "src/gui/captureHandler/ICaptureChangeListener.h"
 
 using kImageAnnotator::KImageAnnotator;
 
@@ -104,6 +105,7 @@ private:
 	DragAndDropHandler *mDragAndDropHandler;
 	UploaderProvider *mUploaderProvider;
 	ICaptureHandler *mCaptureHandler;
+	WidgetHider *mWidgetHider;
 
     void setEnablements(bool enabled);
     void loadSettings();